{"schema":"decodex/commit/1","summary":"Split source capture helper modules","authority":"manual"}
